rag module/requirements.txt
fastapi
uvicorn
openai
pinecone-client
boto3
